+++ NEWS.md
@@ -1,5 +1,13 @@
 # tidyr (development version)
 
+* `chop()` now creates list-columns of class `vctrs::list_of()`. This
+  helps keep track of the type in case the chopped data frame is
+  empty. This allows `unchop()` to reconstitute a data frame with
+  the correct column types even when there are no observations.
+
+* `chop()` now preserves the width of the input data frame even when
+  it has no observation.
+
 * `pivot_longer()` no longer creates a `.copy` variable in the presence of
   duplicate column names. This makes it more consistent with the handling
   of non-unique specs.